Prep for v0.1.0 (#2)
Small tweaks after testing in our internal codebase:

- BestOf doesn't need to be a pointer
- Make pointer float values float64 so literals don't need to be cast (e.g. params.Optional(0.5) vs params.Optional(float32(0.5)))
- Update README.md

+++ edits.go
@@ -23,14 +23,16 @@ type EditsRequest struct {
 	// Temperature specifies what sampling temperature to use. Higher values means the model will take more risks.
 	// Try 0.9 for more creative applications, and 0 (argmax sampling) for ones with a well-defined answer. OpenAI
 	// generally recommends altering this or top_p but not both.
+	//
 	// More on sampling temperature: https://towardsdatascience.com/how-to-sample-from-language-models-682bceb97277
+	//
 	// Defaults to 1.
-	Temperature *float32 `json:"temperature,omitempty"`
+	Temperature *float64 `json:"temperature,omitempty"`
 	// TopP specifies an alternative to sampling with temperature, called nucleus sampling, where the model considers
 	// the results of the tokens with top_p probability mass. So 0.1 means only the tokens comprising the top 10%
 	// probability mass are considered. OpenAI generally recommends altering this or temperature but not both.
 	// Defaults to 1.
-	TopP *float32 `json:"top_p,omitempty"`
+	TopP *float64 `json:"top_p,omitempty"`
 }
 
 // EditsChoice represents one of possible edits.
